fix: #75 Schedule notification
