+++ packages/teacher-app/public/locales/en/translation.json
@@ -218,5 +218,6 @@
   "RECEIVED_FROM_ADMIN": "Received from Admin",
   "RESET_FILTER": "Reset Filter",
   "SELECT_MODULE": "Select Module",
-  "ADD_NEW_NOTIFICATION": "Add New Notification"
+  "ADD_NEW_NOTIFICATION": "Add New Notification",
+  "SEND_LATER": "Send later"
 }
